Renamed WorldExtensions -> GameStateExtensions as that was the class being extended. * Extracted GameStateGenerator interface and implementations from the Client
