+++ src/main/kotlin/gonorth/Client.kt
@@ -2,7 +2,6 @@ package gonorth
 
 import gonorth.domain.*
 import gonorth.slack.toOpt
-import gonorth.world.WorldBuilder
 import kategory.Option
 import kategory.getOrElse
 import java.util.*
@@ -22,7 +21,7 @@ interface GameClient {
 }
 
 
-class SimpleGameClient(var db: Map<String, GameState>, val engine: GoNorth, val worldBuilder: gonorth.GameStateGenerator) : GameClient {
+class SimpleGameClient(var db: Map<String, GameState>, val engine: GoNorth, val worldBuilder: GameStateGenerator) : GameClient {
     override fun takeInput(userId: String, input: String): Option<GameState> {
         // /gnad EAST
         // /gnad DESCRIBE Key
@@ -66,7 +65,7 @@ class SimpleGameClient(var db: Map<String, GameState>, val engine: GoNorth, val 
  *
  * @param goNorth GoNorth game logic
  */
-class TerminalClient(val goNorth: GoNorth, val worldBuilder: gonorth.GameStateGenerator) {
+class TerminalClient(val goNorth: GoNorth, val worldBuilder: GameStateGenerator) {
 
     fun startGame() {
         val player = Player(1000, emptySet(), alive = true)
@@ -123,76 +122,3 @@ class TerminalClient(val goNorth: GoNorth, val worldBuilder: gonorth.GameStateGe
     }
 }
 
-interface GameStateGenerator {
-    fun generate(player: Player, seed: Long): GameState
-}
-
-class SimpleGameStateGenerator : gonorth.GameStateGenerator {
-    override fun generate(player: Player, seed:Long): GameState {
-        val key = Item("Key", "Shiny key, looks useful")
-
-        val p1 = Location(UUID.randomUUID(), "There is a fork in the path.", emptySet())
-        val p2 = Location(UUID.randomUUID(), "The path comes to an abrupt end.", emptySet())
-        val p3 = Location(UUID.randomUUID(), "You went north and died.", emptySet())
-        val p4 = Location(UUID.randomUUID(),
-                "The road continues to the west, whilst a side path heads south.", emptySet())
-        val p5 = Location(UUID.randomUUID(), "A river blocks your path. A key rests on the ground. ", setOf(key))
-        val p6 = Location(UUID.randomUUID(), "To the north you spot a large tower.", emptySet())
-        val p7 = Location(UUID.randomUUID(),
-                "You look at the tower door in front of you. Rocks fall, You die.", emptySet())
-
-
-
-        val world = WorldBuilder().newLocation(p1)
-                .newLocation(p2)
-                .newLocation(p3)
-                .newLocation(p4)
-                .newLocation(p5)
-                .newLocation(p6)
-                .newLocation(p7)
-                .twoWayLink(p1, p2, Move.EAST, Move.WEST,
-                        "You take the path heading east",
-                        "You make your way back to the crossroads")
-                .twoWayLink(p1, p4, Move.WEST, Move.EAST,
-                        "You take the path heading west",
-                        "You make your way back to the crossroads")
-                .twoWayLink(p4, p5, Move.SOUTH, Move.NORTH,
-                        "You take the beaten path south",
-                        "You take the worn path north")
-                .twoWayLink(p4, p6, Move.WEST, Move.EAST,
-                        "You continue along the western trail",
-                        "As you head west, you ponder whether println can print strings")
-                .linkLocation(p1, p3, Move.NORTH, "You stumble ahead")
-                .linkLocation(p6, p7, Move.NORTH, "You head north towards the tower")
-                .world
-
-        val startingText = GameText(
-                "You find yourself lost in a dark forest.",
-                Option.Some("It might be wise to find shelter for the night.")
-        )
-
-        return GameState(startingText, world, p1.id, player, seed)
-    }
-}
-
-class TinyGameStateGenerator : gonorth.GameStateGenerator {
-    override fun generate(player: Player, seed: Long): GameState {
-        val key = Item("Key", "Shiny key, looks useful")
-
-        val p1 = Location(UUID.randomUUID(), "There is a fork in the path. A key rests on the ground", setOf(key))
-        val p3 = Location(UUID.randomUUID(), "You went north and died.", emptySet())
-
-
-        val world = WorldBuilder().newLocation(p1)
-                .newLocation(p3)
-                .linkLocation(p1, p3, Move.NORTH, "You stumble ahead")
-                .world
-
-        val startingText = GameText(
-                "You find yourself lost in a dark forest.",
-                Option.Some("It might be wise to find shelter for the night.")
-        )
-
-        return GameState(startingText, world, p1.id, player, seed)
-    }
-}
